+++ src/module/QuestGenerator.module.ts
